# Conflicts:
#	src/main/java/cn/nukkit/Player.java
#	src/main/java/cn/nukkit/entity/EntityLiving.java

+++ src/main/java/cn/nukkit/item/enchantment/Enchantment.java
@@ -116,7 +116,7 @@ public abstract class Enchantment implements Cloneable, EnchantmentID {
         return this.getMinEnchantAbility(level) + 5;
     }
 
-    public float getProtectionFactor(EntityDamageEvent event) {
+    public int getProtectionFactor(EntityDamageEvent event) {
         return 0;
     }
