+++ src/main/java/cn/nukkit/item/enchantment/protection/EnchantmentProtectionProjectile.java
@@ -25,18 +25,18 @@ public class EnchantmentProtectionProjectile extends EnchantmentProtection {
     }
 
     @Override
-    public double getTypeModifier() {
+    public int getTypeModifier() {
         return 3;
     }
 
     @Override
-    public float getProtectionFactor(EntityDamageEvent e) {
+    public int getProtectionFactor(EntityDamageEvent e) {
         DamageCause cause = e.getCause();
 
         if (level <= 0 || (cause != DamageCause.PROJECTILE)) {
             return 0;
         }
 
-        return (float) (getLevel() * getTypeModifier());
+        return getLevel() * getTypeModifier();
     }
 }
